+++ src/connection/comm5/protocolManager.ts
@@ -1,6 +1,7 @@
 import AirUnpacker from "./airUnpacker";
 import ClientComm5 from "./clientComm5";
 import DataProxy from "../dataProxy";
+import ConversationTarget from "shared/data/conversationTarget";
 
 export default abstract class ProtocolManager {
 	constructor(protected communicationsManager: ClientComm5, protected dataProxy: DataProxy) {
@@ -34,22 +35,22 @@ export default abstract class ProtocolManager {
 	 * Requests a message to be sent to the specified conversation
 	 *
 	 * @param requestID the ID of the request
-	 * @param chatGUID the GUID of the target conversation
+	 * @param conversation the target conversation
 	 * @param message the message to send
 	 * @return whether or not the request was successfully sent
 	 */
-	public abstract sendMessage(requestID: number, chatGUID: string, message: string): boolean;
+	public abstract sendMessage(requestID: number, conversation: ConversationTarget, message: string): boolean;
 	
 	/**
 	 * Sends an attachment file to the specified conversation
 	 *
 	 * @param requestID the ID of the request
-	 * @param chatGUID the GUID of the target conversation
+	 * @param conversation the target conversation
 	 * @param file the file to send
 	 * @param progressCallback a callback called periodically with the number of bytes uploaded
 	 * @return a promise that completes with the file hash once the file has been fully uploaded
 	 */
-	public abstract sendFile(requestID: number, chatGUID: string, file: File, progressCallback: (bytesUploaded: number) => void): Promise<string>;
+	public abstract sendFile(requestID: number, conversation: ConversationTarget, file: File, progressCallback: (bytesUploaded: number) => void): Promise<string>;
 	
 	/**
 	 * Requests the download of a remote attachment
@@ -119,4 +120,40 @@ export default abstract class ProtocolManager {
 	 * @return whether or not the request was successfully sent
 	 */
 	public abstract requestChatCreation(requestID: number, members: string[], service: string): boolean;
+	
+	/**
+	 * Requests the installation of a remote update
+	 * @param updateID The ID of the update to install
+	 * @return Whether the request was successfully sent
+	 */
+	public abstract requestInstallRemoteUpdate(updateID: number): boolean;
+	
+	/**
+	 * Requests a new FaceTime link
+	 * @return Whether the request was successfully sent
+	 */
+	public abstract requestFaceTimeLink(): boolean;
+	
+	/**
+	 * Initiates a new outgoing FaceTime call with the specified addresses
+	 * @param addresses The list of addresses to initiate the call with
+	 * @return Whether the request was successfully sent
+	 */
+	public abstract initiateFaceTimeCall(addresses: string[]): boolean;
+	
+	/**
+	 * Accepts or rejects a pending incoming FaceTime call
+	 * @param caller The name of the caller to accept or reject the call of
+	 * @param accept True to accept the call, or false to reject
+	 * @return Whether the request was successfully sent
+	 */
+	public abstract handleIncomingFaceTimeCall(caller: string, accept: boolean): boolean;
+	
+	/**
+	 * Tells the server to leave the FaceTime call.
+	 * This should be called after the client has connected to the call with the
+	 * FaceTime link to avoid two of the same user connected.
+	 * @return Whether the request was successfully sent
+	 */
+	public abstract dropFaceTimeCallServer(): boolean;
 }