+++ src/resources/text/changelog.md
@@ -1,2 +1,7 @@
-This update improves the security of AirMessage for web,
-and adds compatibility-level support for protocol 5.5
+With AirMessage Server 4.0.0:
+
+- Make and receive FaceTime calls from your browser
+- Check for and install AirMessage Server updates remotely
+- Create new conversations, even on newer versions of macOS
+
+AirMessage for web will also display a warning if you're running a version of AirMessage Server that will soon become unsupported.